Give Truffle recommendations consistent with existing deployment guide
The Truffle migrations page was out of date, incorrectly claiming that
no plugins yet exist which implement a deployment system for Hardhat.
However since that was written, hardhat-deploy has emerged, and
the same text in /guides/deploying.md was already updated accordingly.

So update the paragraph in truffle-migration.md to match, and
cross-link to /guides/deploying.md.

This is relevant to the discussion in #381.