refactor: add execute based on single-graph
The next step of the refactor is to sketch out:

recipes -> recipeGraph -> validation -> transform -> execution graph ->
execute

A skeleton of each of these steps has been put in place, and execution
filled out for basic contract deploys and calls (libs etc still to do.)

To support integration tests a `deploySingleGraph` call has been added
to ignition to mirror the existing `deploy`. The intention is to fill
out functionality and then finally swap `deploySingleGraph` for
`deploy`.