feat: expose config options through deploy
The execution config is now exposed through deploy and wired into the
`hardhat-ignition` plugin config.

Number of blockConfirmations has been added to the config as well

In addition to setting the config from hardhat config we allow the config
to be overriden per `hre.ignition.deploy(module, {config: {...}}` call.
This is needed for tests.

Resolves #312.